requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database dependencies
sqlmodel>=0.0.14
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.7

# Redis and MQTT
redis>=5.0.0
aiomqtt>=2.4.0

# Security and authentication
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
pyotp>=2.9.0

# Configuration and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Rate limiting and monitoring
slowapi>=0.1.9

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
websocket-client>=1.6.4

# Code quality
ruff>=0.1.6
black>=23.11.0 